pyproject.toml
[tool.coverage.run]
omit = ["webhook_server/tests/*"]

[tool.coverage.report]
fail_under = 90
skip_empty = true

[tool.coverage.html]
directory = ".tests_coverage"

[tool.ruff]
preview = true
line-length = 120
fix = true
output-format = "grouped"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "PLC0415", "ARG"]

[tool.ruff.lint.per-file-ignores]
"webhook_server/tests/*" = ["ARG"]

[tool.ruff.format]
exclude = [".git", ".venv", ".mypy_cache", ".tox", "__pycache__"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true
strict_equality = true
extra_checks = true
warn_unused_configs = true
warn_redundant_casts = true

[tool.hatch.build.targets.wheel]
packages = ["webhook_server"]

[project]
name = "github-webhook-server"
version = "3.0.7"
requires-python = ">=3.12"
description = "A webhook server to manage Github repositories and pull requests."
readme = "README.md"
license = "Apache-2.0"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  "build>=1.2.2.post1",
  "colorama>=0.4.6",
  "colorlog>=6.8.2",
  "fastapi>=0.115.0",
  "pygithub>=2.4.0",
  "pyhelper-utils>=0.0.42",
  "pytest-cov>=6.0.0",
  "pytest-mock>=3.14.0",
  "pytest>=8.3.3",
  "python-simple-logger>=1.0.40",
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
  "ruff>=0.6.9",
  "shortuuid>=1.0.13",
  "string-color>=1.2.3",
  "timeout-sampler>=0.0.46",
  "uvicorn[standard]>=0.31.0",
  "httpx>=0.28.1",
  "asyncstdlib>=3.13.1",
  "webcolors>=24.11.1",
  "pyjwt>=2.8.0",
  "pydantic>=2.5.0",
  "psutil>=7.0.0",
  "fastapi-mcp>=0.4.0",
]

[[project.authors]]
name = "Meni Yakove"
email = " myakove@gmail.com"

[[project.authors]]
name = "Ruth Netser"
email = "ruth.netser@gmail.com"

[project.urls]
homepage = "https://github.com/myakove/github-webhook-server"
repository = "https://github.com/myakove/github-webhook-server"
"Bug Tracker" = "https://github.com/myakove/github-webhook-server/issues"

[project.optional-dependencies]
tests = ["pytest-asyncio>=0.26.0", "pytest-xdist>=3.7.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "ipdb>=0.13.13",
  "ipython>=8.12.3",
  "types-colorama>=0.4.15.20240311",
  "types-pyyaml>=6.0.12.20250516",
  "types-requests>=2.32.4.20250611",
]
tests = ["psutil>=7.0.0", "pytest-asyncio>=0.26.0", "pytest-xdist>=3.7.0"]
